+++ react/components/SelectionListOrderBy.js
@@ -3,6 +3,7 @@ import PropTypes from 'prop-types'
 import { injectIntl, intlShape } from 'react-intl'
 import classNames from 'classnames'
 import { find, propEq } from 'ramda'
+import { formatIOMessage } from 'vtex.native-types'
 import { IconCaret } from 'vtex.store-icons'
 import { useDevice } from 'vtex.device-detector'
 import { useCssHandles } from 'vtex.css-handles'
@@ -20,7 +21,7 @@ const CSS_HANDLES = [
   'filterPopupArrowIcon',
 ]
 
-const SelectionListOrderBy = ({ intl, orderBy, options }) => {
+const SelectionListOrderBy = ({ intl, message, orderBy, options }) => {
   const [showDropdown, setShowDropdown] = useState(false)
   const handles = useCssHandles(CSS_HANDLES)
 
@@ -49,6 +50,8 @@ const SelectionListOrderBy = ({ intl, orderBy, options }) => {
       )
     })
   }
+
+  const sortByMessage = formatIOMessage({ id: message, intl })
 
   const getOptionTitle = useCallback(
     option => {
@@ -95,7 +98,7 @@ const SelectionListOrderBy = ({ intl, orderBy, options }) => {
               'dn dib-ns': !orderBy.length,
             })}
           >
-            {intl.formatMessage({ id: 'store/ordenation.sort-by' })}
+            {sortByMessage}
           </span>{' '}
           {getOptionTitle(orderBy)}
         </span>
@@ -123,6 +126,8 @@ SelectionListOrderBy.propTypes = {
   ),
   /** Intl to translations */
   intl: intlShape,
+  /** Message to be displayed */
+  message: PropTypes.string,
 }
 
 export default injectIntl(SelectionListOrderBy)