+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [3.58.0] - 2020-05-12
 ### Added
 - Support for customization of `TotalProductsFlexible` and `OrderByFlexible` components via site-editor.